# Path: optimize `join` by precomputing capacity if possible